requirements.txt
numpy
faiss-cpu
laserembeddings
hydra-core
setuptools